# swagger.yaml
basePath: /
consumes:
- application/json
info:
  description: Documentation of Student API
  title: Student API
  version: 1.0.0
paths:
  /students:
    get:
      description: Returns list of all students
      operationId: listStudents
      responses:
        "200":
          $ref: '#/responses/studentsResponse'
      tags:
      - students
    post:
      description: Create a new student
      operationId: createStudent
      parameters:
      - description: Student data structure to Update or Create.
        in: body
        name: Body
        required: true
        schema:
          properties:
            age:
              format: uint8
              type: integer
              x-go-name: Age
            branch:
              type: string
              x-go-name: Branch
            class:
              type: string
              x-go-name: Class
            name:
              type: string
              x-go-name: Name
          type: object
      responses:
        "200":
          $ref: '#/responses/studentResponse'
      tags:
      - students
  /students/{id}:
    delete:
      description: Returns a student object
      operationId: deleteStudent
      parameters:
      - description: The id of a specific student
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "204":
          $ref: '#/responses/noContent'
      tags:
      - students
    get:
      description: Returns a student object
      operationId: getStudent
      parameters:
      - description: The id of a specific student
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/studentResponse'
      tags:
      - students
    put:
      description: Update a student details
      operationId: updateStudent
      parameters:
      - description: The id of a specific student
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      - description: Student data structure to Update or Create.
        in: body
        name: Body
        required: true
        schema:
          properties:
            age:
              format: uint8
              type: integer
              x-go-name: Age
            branch:
              type: string
              x-go-name: Branch
            class:
              type: string
              x-go-name: Class
            name:
              type: string
              x-go-name: Name
          type: object
      responses:
        "200":
          $ref: '#/responses/studentResponse'
      tags:
      - students
produces:
- application/json
responses:
  noContent:
    description: ""
  studentResponse:
    description: A student object returns in response
    schema:
      properties:
        age:
          format: uint8
          type: integer
          x-go-name: Age
        branch:
          type: string
          x-go-name: Branch
        class:
          type: string
          x-go-name: Class
        id:
          type: string
          x-go-name: Id
        name:
          type: string
          x-go-name: Name
      type: object
  studentsResponse:
    description: A list of students returns in response
    schema:
      items:
        properties:
          age:
            format: uint8
            type: integer
            x-go-name: Age
          branch:
            type: string
            x-go-name: Branch
          class:
            type: string
            x-go-name: Class
          id:
            type: string
            x-go-name: Id
          name:
            type: string
            x-go-name: Name
        type: object
      type: array
schemes:
- http
swagger: "2.0"